First functional version of the Auditor and some extensions
Introduced with the first NBean 'System' Table/Bean all System Beans
should be written in Uppercase to differentiate them from the rest
of all used Beans.

The fluent interface of Bean was enhanced by a new Put method which receives a
Dicionary<string, object>. The Store() and Trash() methods now can also be
used in a fluent manner.

- Introduced new API properties:
 - CurrentUser ... to Audit in the current user's context
 - Copy ... copies/"clones" a Bean
 - GetKeyName()
 - GetNcKeyValue()
 - GetCompoundKeyNames()
 - ToJson()
 - GetDbTypeFromValue()

- Introduced new Bean properties:
 - Columns ... gets all properties
 - Data ... exports the Bean's Dictionary

- Introduced new Bean Methods:
  - GetKeyName()
  - GetKeyValue()
  - GetDirtyBackup()

- New extension method for object: FormatValueToString(), respecting the CurrentCulture.