Relation support for 1:n and m:n relations
Beans can be linked/unlinked in a 1:n or m:n manner, from both sides, based on a referenced or referencing Bean. Lists of owned Beans (1:n) or linked Beans (m:n) can be retreived.
With this Extension to the library the default name for a Primary Key is a lowercased "id" since it's consequently used to build Foreign Key names for relational support.